engine updater pending

+++ src/main/java/com/neo/entity/CustomerEvents.java
@@ -19,10 +19,10 @@ public class CustomerEvents implements Serializable{
     private Long id;
 
     @Column
-    private Integer eventId;
+    private Long eventId;
 
     @Column
-    private Integer customerId;
+    private Long customerId;
 
     @Column
     private Date createdAt;
